+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 0.1.4
+ * IOT-342: Rename strings in streams-layout-storm module
+ * IOT-295: Move classes to streams-runtime-storm
+ * IOT-294: Refactor classes from layout to streams-runtime
+ * IOT-341 Build failing due to unit test failures in storage-atlas module
  * IOT-340: Move Cache to a separate module
  * IOT-315: RulesChannelHandler does not override abstract method setSource
  * IOT-317: Create streams-sdk